✨ Feat : 이미지 최적화 및 컴포넌트 분리
