+++ src/components/common/ProductCard.tsx
@@ -2,6 +2,8 @@
 
 import Link from 'next/link';
 
+import { transformCloudinaryUrl } from '@/utils/common/transformCloudinaryUrl';
+
 import LikeButton from './LikeButton';
 import OptimizedImage from './OptimizedImage';
 
@@ -15,6 +17,7 @@ type ProductCardProps = {
   onToggleLike: () => void;
   scenario?: ProductCardScenario;
   isNameVisible?: boolean;
+  ispriority?: boolean;
 };
 
 const scenarioToClass = (scenario: ProductCardScenario) => {
@@ -60,8 +63,10 @@ const ProductCard = ({
   onToggleLike,
   scenario = 'default',
   isNameVisible = true,
+  ispriority = false,
 }: ProductCardProps) => {
   const classes = scenarioToClass(scenario);
+  const optimizedUrl = transformCloudinaryUrl(imageUrl);
 
   return (
     <div className={classes.container}>
@@ -75,7 +80,12 @@ const ProductCard = ({
         {/* 이미지 영역 */}
         <div className={classes.image}>
           {/* fill 모드 */}
-          <OptimizedImage src={imageUrl} alt={name} fill />
+          <OptimizedImage
+            src={optimizedUrl}
+            alt={name}
+            fill
+            priority={ispriority}
+          />
         </div>
 
         {/* 이름 */}
