+++ src/utils/common/transformCloudinaryUrl.ts
@@ -1,0 +1,3 @@
+export const transformCloudinaryUrl = (originalUrl: string): string => {
+  return originalUrl.replace(/(w_)\d+/, '$1448');
+};
